Document bluebird promisification multiArg option
Add clarity to the README concerning promisification. Just following the existing suggestions in the README leads to a subtle problem for functions that return multiple values. Only the first result is returned without the `multiArg` option. For functions like `addAuthUser` this means the promise resolves to `null` when no mfa is required. The actual response is lost.