Documentation update for .data for axios
The documentation does not include `.data` to access the response data. When using the functions as described in the documentation we are accessing the Axios response object directly so we need to also use `.data` before we access the data items in each response.

This is defined within the typescript but not in the docs.